requirements.txt
python-dotenv~=1.0.1
moexalgo~=2.1.1

numpy~=1.26.4
pandas~=2.2.2

backtrader~=1.9.78.123

python-telegram-bot[job-queue]
requests~=2.31.0

pytz~=2024.1
setuptools~=70.0.0
cachetools~=5.3.3
APScheduler~=3.10.4
pytz>=2018.6
colorama

